Unify source files and add alwaysFalse
