+++ pkg/telemetryquerylanguage/tql/boolean_value.go
@@ -18,6 +18,72 @@ import (
 	"fmt"
 )
 
+type condFunc = func(ctx TransformContext) bool
+
+var alwaysTrue = func(ctx TransformContext) bool {
+	return true
+}
+
+var alwaysFalse = func(ctx TransformContext) bool {
+	return false
+}
+
+// builds a function that returns a short-circuited result of ANDing together condFuncs
+func andFuncs(funcs []condFunc) condFunc {
+	return func(ctx TransformContext) bool {
+		for _, f := range funcs {
+			if !f(ctx) {
+				return false
+			}
+		}
+		return true
+	}
+}
+
+// builds a function that returns a short-circuited result of ORing together condFuncs
+func orFuncs(funcs []condFunc) condFunc {
+	return func(ctx TransformContext) bool {
+		for _, f := range funcs {
+			if f(ctx) {
+				return true
+			}
+		}
+		return false
+	}
+}
+
+func newConditionEvaluator(cond *Condition, functions map[string]interface{}, pathParser PathExpressionParser) (condFunc, error) {
+	if cond == nil {
+		return alwaysTrue, nil
+	}
+	left, err := NewGetter(cond.Left, functions, pathParser)
+	if err != nil {
+		return nil, err
+	}
+	right, err := NewGetter(cond.Right, functions, pathParser)
+	// TODO(anuraaga): Check if both left and right are literals and const-evaluate
+	if err != nil {
+		return nil, err
+	}
+
+	switch cond.Op {
+	case "==":
+		return func(ctx TransformContext) bool {
+			a := left.Get(ctx)
+			b := right.Get(ctx)
+			return a == b
+		}, nil
+	case "!=":
+		return func(ctx TransformContext) bool {
+			a := left.Get(ctx)
+			b := right.Get(ctx)
+			return a != b
+		}, nil
+	}
+
+	return nil, fmt.Errorf("unrecognized boolean operation %v", cond.Op)
+}
+
 func newBooleanExpressionEvaluator(expr *BooleanExpression, functions map[string]interface{}, pathParser PathExpressionParser) (condFunc, error) {
 	if expr == nil {
 		return alwaysTrue, nil
@@ -70,9 +136,10 @@ func newBooleanValueEvaluator(value *BooleanValue, functions map[string]interfac
 		}
 		return condition, nil
 	case value.ConstExpr != nil:
-		return func(ctx TransformContext) bool {
-			return bool(*value.ConstExpr)
-		}, nil
+		if *value.ConstExpr {
+			return alwaysTrue, nil
+		}
+		return alwaysFalse, nil
 	case value.SubExpr != nil:
 		return newBooleanExpressionEvaluator(value.SubExpr, functions, pathParser)
 	}
